mavlink: move ATTITUDE to streams hpp
